Ultroid - v0.0.4
ChangeLog - https://t.me/TheUltroid/26

+++ plugins/autopic.py
@@ -11,6 +11,9 @@
 • `{i}autopic <search query>`
     Will change your profile pic at defined intervals with pics related to the given topic.
 
+• `{i}stoppic`
+    Stop the AutoPic command.
+
 """
 import asyncio
 import os
@@ -25,25 +28,36 @@ from . import *
 async def autopic(e):
     search = e.pattern_match.group(1)
     if not search:
-        return await eor(e, "Heya Give me some Text ..")
+        return await eor(get_string("autopic_1"))
     clls = returnpage(search)
     if len(clls) == 0:
-        return await eor(e, f"No Results found for `{search}`")
+        return await eor(get_string("autopic_2").format(search))
     if not len(clls) == 1:
         num = random.randrange(0, len(clls) - 1)
     else:
         num = 0
     page = clls[num]
     title = page["title"]
-    a = await eor(
-        e, f" Got a Collection `{title}` related to your search !\nStarting Autopic !!"
-    )
+    await eor(get_string("autopic_3").format(title))
+    udB.set("AUTOPIC", "True")
     while True:
+        ge = udB.get("AUTOPIC")
+        if not ge == "True":
+            return
         animepp(page["href"])
         file = await ultroid_bot.upload_file("autopic.jpg")
         await ultroid_bot(functions.photos.UploadProfilePhotoRequest(file))
         os.remove("autopic.jpg")
         await asyncio.sleep(1100)
+
+
+@ultroid_cmd(pattern="stoppic$")
+async def stoppo(ult):
+    gt = udB.get("AUTOPIC")
+    if not gt == "True":
+        return await eor(ult, "`AUTOPIC was not in used !!`")
+    udB.set("AUTOPIC", "None")
+    await eor(ult, "`AUTOPIC Stopped !!`")
 
 
 HELP.update({f"{__name__.split('.')[1]}": f"{__doc__.format(i=HNDLR)}"})
